1.15.15 refactor: capabilities to own package
